Move the component parts of the charts into their own directory
