+++ lib/prawn-graph.rb
@@ -3,7 +3,8 @@ require "prawn"
 require "prawn/graph/version"
 require "prawn/graph/calculations"
 
-require "prawn/graph/canvas"
+require "prawn/graph/chart_components"
+
 require "prawn/graph/theme"
 require "prawn/graph/series"
 require "prawn/graph/charts"
