+++ lib/prawn/graph/chart_components.rb
@@ -1,0 +1,1 @@
+require_relative "chart_components/canvas"
